refactor(flutter_client): ディレクトリ構成の変更
