+++ flutter_client/lib/model/entities/project.dart
