+++ flutter_client/lib/model/types/color_palette_type.dart
