+++ flutter_client/lib/model/types/knitting_pattern_size.dart
