+++ flutter_client/lib/model/types/knitting_type.dart
